+++ test/integration/shared/sdk-outgoing-requests.test.ts
@@ -153,7 +153,7 @@ describe('SDKOutgoingRequests', () => {
 
   describe('requestTransfer', () => {
     it('should return transfer details', async () => {
-      const transferId = uuid();
+      const transferId = uuid()
       const request: OutboundAPI.Schemas.simpleTransfersPostRequest = {
         fspId: 'dfspa',
         transfersPostRequest: {